requirements.txt
black==23.3.0
isort==5.12.0
langchain==0.0.167
numpy==1.24.3
openai==0.27.6
pandas==2.0.1
plotly==5.14.1
tweety-ns==0.7.1